+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2019-07-28
+
+### Added
+- Provide support for direct-path loading with Oracle DB (ER #8)
+- Provide support for custom quote characters (ER #10)
+
 ## [1.2.0] - 2019-03-09
 
 ### Added